refactor : refactor test package
