+++ src/test/java/lotto/model/domain/verifier/InputVerifierTest.java
@@ -1,7 +1,6 @@
-package lotto.model.domain;
+package lotto.model.domain.verifier;
 
 
-import lotto.model.domain.verifier.InputVerifier;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
